fixe: duplicated files

+++ backend/src/main/java/lk/fat2fit/Fat2Fit/Repository/ClientBodyMetricsRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface ClientBodyMetricsRepository extends JpaRepository<ClientBodyMetrics, Integer> {
 
-    Optional<ClientBodyMetrics> findByClientId(int clientId);
+    Optional<ClientBodyMetrics> findByClientId(Long clientId);
 }
